Use getProps method to dynamically watch for changes in props inside Tasks.
Pass a getProps method as the second argument to a Task's generator
which returns a promise that resolves with the Task's current props as
soon as the given filter function returns true (or resolves immediately
with the current props if no filter function is provided).

This makes it easier to pass information about the current state of the
application  to a Task while it is in the middle of execution.

Also, get rid of the warning about ambiguity about whether tasks should
be started/stopped now that there is a built in way to pass information
to the Task during execution.