Reimplement proc in a functional style.
The new implementation has two main functions: runProc and stepProc.

stepProc executes one step of the proc and returns a promise that
resolves when the step is done. This is a low level function that other
abstractions could potentially be built on top of (such as a
pause/resumable proc, for example).

runProc runs stepProc in a loop for you, and returns a promise that
resolves when the generator function given is done executing.