Remove run() method from Task and don't pass initial props to generator.
Instead of making subclasses of the Task class that override the run
method, now you have to use the Task component and specify the
`generator` prop.

Also, the generator function does not get passed the initial props as
the first argument anymore. Instead, it only gets passed the getProps
function, which should be used to get the initial props.

For example:

```
function* someTask(getProps) {
  const initialProps = yield call(getProps);
}
```

This might be a bit of a pain, but this makes it much more clear that
it is just the *initial* props, not the current props. To get the
current props, you have to call getProps again.