Add tests for stepProc/runProc and several small but breaking changes.
* Add tests for `stepProc`/`runProc`.
* The `cancel` method of the promise returned by stepProc no longer
cancels the execution of the entire proc, just cancels the current step.
* Instead, use the new `stopProc()` function to stop the execution of a
proc.
* Change `runProc` to take a generator instead of a generator function to
keep it consistent with `stepProc`/`stopProc`.
* Add a new `run()` convenience function that takes a generator function
and arguments to be passed to it.
* Add an `onCall` callback to stepProc and runProc, which gets called
whenever a call object gets executed with the call object and the result
of the call. This is not meant to be a public API, it's just a hack I
threw in to make it possible to see what the last call for a given Task
within the React dev tools.