Add synchronous version of stepProc and remove experimental onCall callback.
Add `stepProcSynchronous` which immediately returns the result of a
step, and add a new step type called RESULT_TYPE_WAIT to indicate that
the execution loop that uses stepProcSynchronous should wait on the
promise contained in the value field.

This is used in `runProc` to add an onStep callback, which replaces
onCall, which can be used to see what the currently executing step is.

This is used to add the current step to the Task component's state so it
can be inspected with the React dev tools.