Fix promise cancellation and remove cancel() method from async procs.
Cancellable promises will not resolve or reject after your call their
cancel() method, but asynchronous procs need to be able to keep on
running to finish executing any remaining `try { } finally { }` blocks.

Because the semantics of cancellation are different for procs, replace
the cancel() method returned by procs with a return() method, which
calls the return() method on the generator and continues execution, and
a stop() method, which just stops execution and immediately resolves.