Refactor Task component and task() helper.
task() now returns a element task element with the proc and key set for
you, instead of a new component that renders a Task. You can still make
wrapper components for your Tasks if you want to, but you don't have to.

Update Task to use the new proc API and to log all of the calls and
results in non-production environments, as well as warn you if a Task's
proc doesn't stop within 10 seconds of being unmounted.

Also update the basic counter example to use the new Task/proc.